more permissions for admin

+++ backend/app1/Views/CoachViews.py
@@ -5,7 +5,7 @@ from rest_framework import status, generics
 
 from .Pagination import CustomPagination
 from ..models import Coach, Team
-from ..permissions import HasEditPermissionOrReadOnly
+from ..permissions import HasEditPermissionOrReadOnly, IsAdminOrReadOnly
 from ..serailizer import CoachSerializer, CoachSerializerId
 
 
@@ -92,7 +92,7 @@ class CoachInfo(APIView):
 
 
 class CoachBulk(APIView):
-    # permission_classes = [IsAuthenticatedOrReadOnly, HasEditPermissionOrReadOnly]
+    permission_classes = [IsAdminOrReadOnly]
 
     def delete(self, request, *args, **kwargs):
         ids = kwargs.get('ids')
@@ -100,7 +100,7 @@ class CoachBulk(APIView):
         if ids:
             ids_list = ids.split(',')
             queryset = Coach.objects.filter(id__in=ids_list)
-            # self.check_object_permissions(request, queryset)
+            self.check_object_permissions(request, queryset)
             deleted_count, _ = queryset.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
