+++ backend/app1/Views/FanViews.py
@@ -6,7 +6,7 @@ from rest_framework import status, generics
 
 from .Pagination import CustomPagination
 from ..models import Fan, SwimmerFan
-from ..permissions import HasEditPermissionOrReadOnly
+from ..permissions import HasEditPermissionOrReadOnly, IsAdminOrReadOnly
 from ..serailizer import FanSerializer, FanSerializerId, SwimmerFanSerializer, FanSerializerAvg
 
 
@@ -134,7 +134,7 @@ class FansOrderedByName(generics.ListCreateAPIView):
 
 
 class FanBulk(APIView):
-    # permission_classes = [IsAuthenticatedOrReadOnly, HasEditPermissionOrReadOnly]
+    permission_classes = [IsAdminOrReadOnly]
 
     def delete(self, request, *args, **kwargs):
         ids = kwargs.get('ids')
@@ -142,7 +142,7 @@ class FanBulk(APIView):
         if ids:
             ids_list = ids.split(',')
             queryset = Fan.objects.filter(id__in=ids_list)
-            # self.check_object_permissions(request, queryset)
+            self.check_object_permissions(request, queryset)
             deleted_count, _ = queryset.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
