+++ backend/app1/Views/SwimmerFanViews.py
@@ -5,7 +5,7 @@ from rest_framework import status, generics
 
 from .Pagination import CustomPagination
 from ..models import SwimmerFan
-from ..permissions import HasEditPermissionOrReadOnly
+from ..permissions import HasEditPermissionOrReadOnly, IsAdminOrReadOnly
 from ..serailizer import SwimmerFanSerializer
 
 
@@ -85,7 +85,7 @@ class SwimmerFanInfo(APIView):
 
 
 class SwimmerFanBulk(APIView):
-    # permission_classes = [IsAuthenticatedOrReadOnly, HasEditPermissionOrReadOnly]
+    permission_classes = [IsAdminOrReadOnly]
 
     def delete(self, request, *args, **kwargs):
         ids = kwargs.get('ids')
@@ -93,7 +93,7 @@ class SwimmerFanBulk(APIView):
         if ids:
             ids_list = ids.split(',')
             queryset = SwimmerFan.objects.filter(id__in=ids_list)
-            # self.check_object_permissions(request, queryset)
+            self.check_object_permissions(request, queryset)
             deleted_count, _ = queryset.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
