+++ backend/app1/Views/TeamViews.py
@@ -6,7 +6,7 @@ from rest_framework import status, generics
 
 from .Pagination import CustomPagination
 from ..models import Team, Swimmer
-from ..permissions import HasEditPermissionOrReadOnly
+from ..permissions import HasEditPermissionOrReadOnly, IsAdminOrReadOnly
 from ..serailizer import TeamSerializer, TeamSerializerNo, SwimmerSerializer
 
 
@@ -157,7 +157,7 @@ class TeamsOrderedByName(generics.ListCreateAPIView):
 
 
 class TeamBulk(APIView):
-    # permission_classes = [IsAuthenticatedOrReadOnly, HasEditPermissionOrReadOnly]
+    permission_classes = [IsAdminOrReadOnly]
 
     def delete(self, request, *args, **kwargs):
         ids = kwargs.get('ids')
@@ -165,7 +165,7 @@ class TeamBulk(APIView):
         if ids:
             ids_list = ids.split(',')
             queryset = Team.objects.filter(id__in=ids_list)
-            # self.check_object_permissions(request, queryset)
+            self.check_object_permissions(request, queryset)
             deleted_count, _ = queryset.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
